+++ pkg/output/options.go
@@ -7,13 +7,14 @@ package output
 import "time"
 
 type Options struct {
-	Addr        string        // Destination address (host:port).
-	Delay       time.Duration // Delay start after start signal.
-	Protocol    string        // Protocol (udp/tcp/tls).
-	Retries     int           // Number of connection retries for tcp based protocols.
-	StartSignal string        // OS signal to wait on before starting.
-	InsecureTLS bool          // Disable TLS verification checks.
-	RateLimit   int           // UDP rate limit in bytes.
+	Addr            string        // Destination address (host:port).
+	Delay           time.Duration // Delay start after start signal.
+	Protocol        string        // Protocol (udp/tcp/tls).
+	Retries         int           // Number of connection retries for tcp based protocols.
+	StartSignal     string        // OS signal to wait on before starting.
+	InsecureTLS     bool          // Disable TLS verification checks.
+	RateLimit       int           // UDP rate limit in bytes.
+	LogReaderBuffer int           // Log reader buffer size in bytes.
 
 	WebhookOptions
 	GCPPubsubOptions