+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- Added option to set up custom buffer size for the log reader. [#22](https://github.com/elastic/stream/pull/22)
+- Added support for glob patterns. [#22](https://github.com/elastic/stream/pull/22)
+
 ## [0.4.0]
 
 - Added HTTP Server output. [#19](https://github.com/elastic/stream/pull/19)